Sobrecarga de métodos e construtodres

+++ src/com/maratorajava/javacore/introducaometodos/school/classes/GradeCalculator.java
@@ -8,7 +8,7 @@ public class GradeCalculator {
             return 0;
         }
 
-        int total = 0;
+        float total = 0;
         int countGrades = studentGrades.length;
 
         for(Grade grade : studentGrades) {
